Remove `useSyncedQueryString`
